backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
requests
openai-whisper
ffmpeg-python